backend/pyproject.toml
[project]
name = "nutribot-backend"
version = "0.1.0"
description = "NutriBot - AI-powered nutrition assistant backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "pydantic==2.10.4",
    "pydantic-settings==2.6.1",
    "openai==1.58.1",
    "tenacity==9.0.0",
    "requests==2.32.3",
    "SQLAlchemy==2.0.36",
    "alembic==1.14.0",
    "psycopg2-binary==2.9.10",
    "Pillow>=10.0.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "pyright>=1.1.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
